+++ CHANGELOG.md
@@ -109,6 +109,7 @@
 - Chg #999: Remove `requireTransaction()` method and `$isolationLevel` property from `AbstractCommand` (@vjik)
 - Enh #1000: Prepare values in `HashCondition` (@vjik)
 - Chg #1001: Remove `ParamInterface` (@vjik)
+- Chg #1001: Add public properties `$type` and `$value` to `Param` class instead of `getType()` and `getValue()` methods that were removed (@vjik)
 
 ## 1.3.0 March 21, 2024